requirements.txt
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
scikit-image==0.22.0
opencv-python-headless==4.9.0.80
tqdm==4.66.2
matplotlib==3.8.3
tensorboard==2.16.2
pyyaml==6.0.1
ftfy==6.2.0
regex==2023.12.25
ultralytics==8.0.225